voiceai-agent/requirements.txt
livekit-agents[elevenlabs,silero]>=0.8.0
livekit>=0.17.0
elevenlabs>=1.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
